refactor: move main.ts to src folder
