+++ tsconfig.json
@@ -19,6 +19,6 @@
     ]
   },
   "include": [
-    "**/*.ts"
+    "src/**/*.ts"
   ]
 }
